[core] Fix behavior of multiple char update packets in single tick
* After the 0x00D/0x037 overhaul, the logic to update a char update packet broke. This fixes it.
